remove all old classes and refactor current classes into root directories -- final step before merging
