+++ android/app/src/main/java/io/o2mc/app/App.java
@@ -2,15 +2,15 @@ package io.o2mc.app;
 
 import android.app.Application;
 
-import io.o2mc.sdk.current.O2MC;
+import io.o2mc.sdk.O2MC;
 
-public class AppNew extends Application {
+public class App extends Application {
 
-    private static final String TAG = "AppNew";
+    private static final String TAG = "App";
 
     private static O2MC o2mc;
 
-    public AppNew() {
+    public App() {
 //        o2mc = new O2MC(null, null);
         o2mc = new O2MC(this, "http://10.0.2.2:5000/events");
         o2mc.setDispatchInterval(8);
