+++ android/sdk/src/main/java/io/o2mc/sdk/business/EventBus.java
@@ -1,11 +1,11 @@
-package io.o2mc.sdk.current.business;
+package io.o2mc.sdk.business;
 
 import android.util.Log;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import io.o2mc.sdk.current.domain.Event;
+import io.o2mc.sdk.domain.Event;
 
 /**
  * Holds all events generated by user interaction from the app implementing our SDK.
