+++ android/sdk/src/main/java/io/o2mc/sdk/domain/Batch.java
@@ -1,4 +1,4 @@
-package io.o2mc.sdk.current.domain;
+package io.o2mc.sdk.domain;
 
 import java.util.List;
