--- android/sdk/src/main/java/io/o2mc/sdk/old/datastreams/Connectivity.java
@@ -1,218 +1,0 @@
-package io.o2mc.sdk.old.datastreams;
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-import android.telephony.TelephonyManager;
-
-import java.net.InetAddress;
-import java.net.InterfaceAddress;
-import java.net.NetworkInterface;
-import java.net.SocketException;
-import java.util.Enumeration;
-
-/**
- * Class pulled from Github. (With some small changes)
- * It returns the current networkstate as a String
- */
-@SuppressWarnings("WeakerAccess")
-public class Connectivity {
-
-    /**
-     * Get the network info
-     *
-     * @param context
-     * @return null if unable to retrieve active network information
-     */
-    @SuppressWarnings("JavaDoc")
-    public static NetworkInfo getNetworkInfo(Context context) {
-        ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-        //noinspection ConstantConditions -- null is being dealt with in calling methods
-        return cm.getActiveNetworkInfo();
-    }
-
-    /**
-     * Check if there is any connectivity
-     *
-     * @param context
-     * @return
-     */
-    @SuppressWarnings("JavaDoc")
-    public static boolean isConnected(Context context) {
-        NetworkInfo info = Connectivity.getNetworkInfo(context);
-        return (info != null && info.isConnected());
-    }
-
-    /**
-     * Check if there is any connectivity to a Wifi network
-     *
-     * @param context
-     * @return
-     */
-    @SuppressWarnings("JavaDoc")
-    public static boolean isConnectedWifi(Context context) {
-        NetworkInfo info = Connectivity.getNetworkInfo(context);
-        return (info != null && info.isConnected() && info.getType() == ConnectivityManager.TYPE_WIFI);
-    }
-
-    /**
-     * Check if there is any connectivity to a mobile network
-     *
-     * @param context
-     * @return
-     */
-    @SuppressWarnings("JavaDoc")
-    public static boolean isConnectedMobile(Context context) {
-        NetworkInfo info = Connectivity.getNetworkInfo(context);
-        return (info != null && info.isConnected() && info.getType() == ConnectivityManager.TYPE_MOBILE);
-    }
-
-    /**
-     * Check if there is fast connectivity
-     *
-     * @param context
-     * @return true means the user has a fast network connection
-     */
-    @SuppressWarnings("JavaDoc")
-    public static boolean isConnectedFast(Context context) {
-        NetworkInfo info = Connectivity.getNetworkInfo(context);
-        return (info != null && info.isConnected() && Connectivity.isConnectionFast(info.getType(), info.getSubtype()));
-    }
-
-    /**
-     * Returns the IP address of the network to which the user's phone is currently connected to.
-     *
-     * @return InetAddress object if connected to the internet, null if not
-     * @throws SocketException on failure to acquire network interfaces or executing operations on them
-     */
-    static InetAddress ip() throws SocketException {
-        Enumeration<NetworkInterface> nis = NetworkInterface.getNetworkInterfaces();
-        NetworkInterface ni;
-        while (nis.hasMoreElements()) {
-            ni = nis.nextElement();
-            if (!ni.isLoopback()/*not loopback*/ && ni.isUp()/*it works now*/) {
-                for (InterfaceAddress ia : ni.getInterfaceAddresses()) {
-                    //filter for ipv4/ipv6
-                    if (ia.getAddress().getAddress().length == 4) {
-                        //4 for ipv4, 16 for ipv6
-                        return ia.getAddress();
-                    }
-                }
-            }
-        }
-        return null;
-    }
-
-    public static String getConnectivityType(Context context) {
-
-        NetworkInfo info = Connectivity.getNetworkInfo(context);
-        int type = info.getType();
-        int subType = info.getSubtype();
-
-        switch (type) {
-            case ConnectivityManager.TYPE_WIFI:
-                return "WIFI";
-            case ConnectivityManager.TYPE_MOBILE:
-                switch (subType) {
-                    case TelephonyManager.NETWORK_TYPE_1xRTT:
-                        return "1xRTT"; // ~ 50-100 kbps
-                    case TelephonyManager.NETWORK_TYPE_CDMA:
-                        return "CDMA"; // ~ 14-64 kbps
-                    case TelephonyManager.NETWORK_TYPE_EDGE:
-                        return "EDGE"; // ~ 50-100 kbps
-                    case TelephonyManager.NETWORK_TYPE_EVDO_0:
-                        return "EVDO_0"; // ~ 400-1000 kbps
-                    case TelephonyManager.NETWORK_TYPE_EVDO_A:
-                        return "EVDO_A"; // ~ 600-1400 kbps
-                    case TelephonyManager.NETWORK_TYPE_GPRS:
-                        return "GPRS"; // ~ 100 kbps
-                    case TelephonyManager.NETWORK_TYPE_HSDPA:
-                        return "HSDPA"; // ~ 2-14 Mbps
-                    case TelephonyManager.NETWORK_TYPE_HSPA:
-                        return "HSPA"; // ~ 700-1700 kbps
-                    case TelephonyManager.NETWORK_TYPE_HSUPA:
-                        return "HSUPA"; // ~ 1-23 Mbps
-                    case TelephonyManager.NETWORK_TYPE_UMTS:
-                        return "UMTS"; // ~ 400-7000 kbps
-                    /*
-                     * Above API level 7, make sure to set android:targetSdkVersion
-                     * to appropriate level to use these
-                     */
-                    case TelephonyManager.NETWORK_TYPE_EHRPD: // API level 11
-                        return "EHRPD"; // ~ 1-2 Mbps
-                    case TelephonyManager.NETWORK_TYPE_EVDO_B: // API level 9
-                        return "EVDO_B"; // ~ 5 Mbps
-                    case TelephonyManager.NETWORK_TYPE_HSPAP: // API level 13
-                        return "HSPAP"; // ~ 10-20 Mbps
-                    case TelephonyManager.NETWORK_TYPE_IDEN: // API level 8
-                        return "IDEN"; // ~25 kbps
-                    case TelephonyManager.NETWORK_TYPE_LTE: // API level 11
-                        return "LTE"; // ~ 10+ Mbps
-                    // Unknown
-                    case TelephonyManager.NETWORK_TYPE_UNKNOWN:
-                    default:
-                        return "unknown";
-                }
-            default:
-                return "unknown";
-        }
-    }
-
-    /**
-     * Check if the connection is fast
-     *
-     * @param type    one of the static constants as defined in ConnectivityManager
-     * @param subType if type was of TYPE_MOBILE, subType must be one of TelephonyManager's constants
-     * @return true means the user has a fast network connection
-     */
-    public static boolean isConnectionFast(int type, int subType) {
-        switch (type) {
-            case ConnectivityManager.TYPE_WIFI:
-                return true;
-            case ConnectivityManager.TYPE_MOBILE:
-                switch (subType) {
-                    case TelephonyManager.NETWORK_TYPE_1xRTT:
-                        return false; // ~ 50-100 kbps
-                    case TelephonyManager.NETWORK_TYPE_CDMA:
-                        return false; // ~ 14-64 kbps
-                    case TelephonyManager.NETWORK_TYPE_EDGE:
-                        return false; // ~ 50-100 kbps
-                    case TelephonyManager.NETWORK_TYPE_EVDO_0:
-                        return true; // ~ 400-1000 kbps
-                    case TelephonyManager.NETWORK_TYPE_EVDO_A:
-                        return true; // ~ 600-1400 kbps
-                    case TelephonyManager.NETWORK_TYPE_GPRS:
-                        return false; // ~ 100 kbps
-                    case TelephonyManager.NETWORK_TYPE_HSDPA:
-                        return true; // ~ 2-14 Mbps
-                    case TelephonyManager.NETWORK_TYPE_HSPA:
-                        return true; // ~ 700-1700 kbps
-                    case TelephonyManager.NETWORK_TYPE_HSUPA:
-                        return true; // ~ 1-23 Mbps
-                    case TelephonyManager.NETWORK_TYPE_UMTS:
-                        return true; // ~ 400-7000 kbps
-                    /*
-                     * Above API level 7, make sure to set android:targetSdkVersion
-                     * to appropriate level to use these
-                     */
-                    case TelephonyManager.NETWORK_TYPE_EHRPD: // API level 11
-                        return true; // ~ 1-2 Mbps
-                    case TelephonyManager.NETWORK_TYPE_EVDO_B: // API level 9
-                        return true; // ~ 5 Mbps
-                    case TelephonyManager.NETWORK_TYPE_HSPAP: // API level 13
-                        return true; // ~ 10-20 Mbps
-                    case TelephonyManager.NETWORK_TYPE_IDEN: // API level 8
-                        return false; // ~25 kbps
-                    case TelephonyManager.NETWORK_TYPE_LTE: // API level 11
-                        return true; // ~ 10+ Mbps
-                    // Unknown
-                    case TelephonyManager.NETWORK_TYPE_UNKNOWN:
-                    default:
-                        return false;
-                }
-            default:
-                return false;
-        }
-    }
-
-}
