+++ android/sdk/src/test/java/io/o2mc/sdk/business/UtilTest.java
@@ -1,4 +1,4 @@
-package io.o2mc.sdk.current.business;
+package io.o2mc.sdk.business;
 
 import org.junit.After;
 import org.junit.Before;
